@@ -1,7 +1,8 @@
 package com.dev.customerservice.infraestructure.data.entities;
 
 
-import javax.persistence.*;
+
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
