@@ -1,20 +1,17 @@
 package com.dev.customerservice.application.api;
 
+import com.dev.customerservice.application.services.ClienteApplicationService;
 import com.dev.customerservice.domain.core.model.ClienteDto;
 import com.dev.customerservice.domain.core.model.ReporteClienteMovimientosDto;
-import com.dev.customerservice.domain.core.ClienteServicio;
+import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
 import java.time.LocalDate;
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Slf4j
 @RestController
@@ -22,13 +19,13 @@ import java.util.stream.Collectors;
 public class ClienteApi {
 
     @Autowired
-    private ClienteServicio clienteServicio;
+    private ClienteApplicationService clienteApplicationService;
 
 
     @PostMapping
     public ResponseEntity<?> crearCliente(@Valid @RequestBody ClienteDto clienteDto) {
 
-        return ResponseEntity.ok(clienteServicio.crearCliente(clienteDto));
+        return ResponseEntity.ok(clienteApplicationService.crearCliente(clienteDto));
 
 
     }
@@ -37,7 +34,7 @@ public class ClienteApi {
     public ResponseEntity<ReporteClienteMovimientosDto> movimientos(@RequestParam(required = false) @DateTimeFormat(pattern = "yyyy-MM-dd") final LocalDate createDateFrom,
                                                                     @RequestParam(required = false) @DateTimeFormat(pattern = "yyyy-MM-dd") final LocalDate createDateTo,
                                                                     String identificacion) {
-        return new ResponseEntity<>(clienteServicio.reporteClienteMovimientos(createDateFrom, createDateTo, identificacion), HttpStatus.OK);
+        return new ResponseEntity<>(clienteApplicationService.reporteClienteMovimientos(createDateFrom, createDateTo, identificacion), HttpStatus.OK);
     }
 
 }
