@@ -2,6 +2,8 @@ package com.dev.customerservice.domain.core;
 
 import com.dev.customerservice.domain.core.mapper.ClienteMapper;
 import com.dev.customerservice.domain.core.model.*;
+import com.dev.customerservice.domain.ports.ClientePort;
+import com.dev.customerservice.exceptions.PersonaNoDisponibleException;
 import com.dev.customerservice.infraestructure.data.entities.Cliente;
 import com.dev.customerservice.infraestructure.data.entities.Persona;
 import com.dev.customerservice.infraestructure.data.repository.ClienteRepository;
@@ -16,9 +18,10 @@ import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
-public class ClienteServicio {
+public class ClienteServicio implements ClientePort {
     private static final Logger log = LoggerFactory.getLogger(ClienteServicio.class);
     @Autowired
     private ClienteRepository clienteRepository;
@@ -38,7 +41,7 @@ public class ClienteServicio {
     @Autowired
     private ClienteMapper clienteMapper;
 
-
+    @Override
     public ClienteDto crearCliente(ClienteDto clienteDto) {
 
         Cliente cliente = clienteRepository.save(clienteMapper.convertToCliente(clienteDto));
@@ -53,26 +56,20 @@ public class ClienteServicio {
         return clienteMapper.convertToClienteDto(cliente);
     }
 
-
+    @Override
     public ReporteClienteMovimientosDto reporteClienteMovimientos(LocalDate createDateFrom,
                                                                   LocalDate createDateTo,
                                                                   String identificacion) {
 
         List<ReporteClienteDto> cuentaDto = new ArrayList<>();
-        Persona persona = personaRepository.findByIdentificacion(identificacion);
+        Optional<Persona> personaOptional = personaRepository.findByIdentificacion(identificacion);
 
+        if (!personaOptional.isPresent()) {
+            throw new PersonaNoDisponibleException("Persona no disponible");
+
+        }
+        Persona persona = personaOptional.get();
         List<Cuenta> cuentas = cuentaServicio.findByClienteId(((Cliente) persona).getClienteId());
-
-//        cuentas.forEach(cuenta -> {
-//            Movimientos movimientos = movimientoServicio.findByCuentaFecha(cuenta.getId()
-//                    , createDateFrom.format(DateTimeFormatter.ISO_DATE)
-//                    , createDateTo.format(DateTimeFormatter.ISO_DATE));
-//            ReporteClienteDto reporteClienteDto = new ReporteClienteDto();
-//            reporteClienteDto.setNumeroCuenta(cuenta.getNumeroCuenta());
-//            reporteClienteDto.setTipoCuenta(cuenta.getTipoCuenta());
-//            reporteClienteDto.setMovimientos(movimientos);
-//            cuentaDto.add(reporteClienteDto);
-//        });
 
         for (Cuenta cuenta : cuentas) {
             Movimientos movimientos = movimientoServicio.findByCuentaFecha(cuenta.getId()
