@@ -1,48 +1,51 @@
 package com.dev.customerservice.domain.core;
 
+
 import com.dev.customerservice.domain.core.model.EmbeddedDto;
 import com.dev.customerservice.domain.core.model.Movimiento;
 import com.dev.customerservice.domain.core.model.Movimientos;
-import org.springframework.core.ParameterizedTypeReference;
+import com.dev.customerservice.infraestructure.adapters.feign.MovimientoFeignClient;
+import feign.FeignException;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-
 @Service
-public class MovimientoServicio extends RestTemplateAbstract {
+public class MovimientoServicio {
 
-    public Movimiento findByCuenta(Long cuentaId){
-        try{
-            ResponseEntity<EmbeddedDto<Movimientos>> response = findByParameters(
-                    urlMovimientos + "/search/findByCuenta?cuentaId={cuentaId}",
-                    new ParameterizedTypeReference<EmbeddedDto<Movimientos>>() {
-                    },cuentaId);
+    private final MovimientoFeignClient movimientoFeignClient;
 
-            if (response.getBody() != null) {
+    @Autowired
+    public MovimientoServicio(MovimientoFeignClient movimientoFeignClient) {
+        this.movimientoFeignClient = movimientoFeignClient;
+    }
+
+    public Movimiento findByCuenta(Long cuentaId) {
+        try {
+            ResponseEntity<EmbeddedDto<Movimientos>> response = movimientoFeignClient.findByCuenta(cuentaId);
+
+            if (response.getBody() != null && response.getBody().getEmbedded() != null &&
+                    !response.getBody().getEmbedded().getMovimientos().isEmpty()) {
                 return response.getBody().getEmbedded().getMovimientos().get(0);
             }
-
-        }catch (Exception e){
+        } catch (FeignException e) {
+            // Manejar excepciones específicas de Feign
             e.printStackTrace();
         }
         return new Movimiento();
     }
 
-    public Movimientos findByCuentaFecha(Long cuentaId , String createDateFrom,
-                                              String createDateTo){
-        try{
-            ResponseEntity<EmbeddedDto<Movimientos>> response = findByParameters(
-                    urlMovimientos + "/search/findByCuentaFecha?cuentaId={cuentaId}&createDateFrom={createDateFrom}&createDateTo={createDateTo}",
-                   new ParameterizedTypeReference<EmbeddedDto<Movimientos>>() {
-                    },cuentaId,createDateFrom,createDateTo);
+    public Movimientos findByCuentaFecha(Long cuentaId, String createDateFrom, String createDateTo) {
+        try {
+            ResponseEntity<EmbeddedDto<Movimientos>> response = movimientoFeignClient.findByCuentaFecha(
+                    cuentaId, createDateFrom, createDateTo
+            );
 
             if (response.getBody() != null) {
                 return response.getBody().getEmbedded();
             }
-
-        }catch (Exception e){
+        } catch (FeignException e) {
+            // Manejar excepciones específicas de Feign
             e.printStackTrace();
         }
         return null;
