@@ -1,7 +1,6 @@
 package com.dev.customerservice.infraestructure.data.entities;
-import javax.persistence.*;
 
-import lombok.Builder;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
